+++ src/common/index.ts
@@ -4,5 +4,6 @@ export * from './database';
 export * from './decorator';
 export * from './message';
 export * from './swagger';
+export * from './template';
 export * from './type';
 export * from './util';
