+++ src/common/util/index.ts
@@ -2,4 +2,3 @@ export * from './error-util';
 export * from './env-util';
 export * from './date-util';
 export * from './util';
-export * from './slack';
